complete multipartUpload

+++ backend/FileService/FileService.Core/Features/StartMultiPartUploadResponse.cs
@@ -6,4 +6,4 @@ public record StartMultiPartUploadResponse(
     Guid MediaAssetId,
     string UploadId,
     IReadOnlyList<ChunkUploadUrl> ChunkUploadUrls,
-    long ChunkSize);
+    int ChunkSize);
