+++ backend/FileService/FileService.Core/FilesStorage/IChunkSizeCalculator.cs
@@ -5,6 +5,6 @@ namespace FileService.Core.FilesStorage
 {
     public interface IChunkSizeCalculator
     {
-        Result<(long ChunkSize, int TotalChunks), Error> CalculateChunkSize(long fileSize);
+        Result<(int ChunkSize, int TotalChunks), Error> CalculateChunkSize(long fileSize);
     }
 }
