+++ backend/FileService/FileService.Infrastructure.S3/ChunkSizeCalculator.cs
@@ -14,13 +14,13 @@ namespace FileService.Infrastructure.S3
             _options = options.Value;
         }
 
-        public Result<(long ChunkSize, int TotalChunks), Error> CalculateChunkSize(long fileSize)
+        public Result<(int ChunkSize, int TotalChunks), Error> CalculateChunkSize(long fileSize)
         {
             if (_options.RecommendedChunkSizeBytes <= 0 || _options.MaxChunks <= 0)
                 return GeneralErrors.ValueIsInvalid("настройки чанков");
 
             if (fileSize <= _options.RecommendedChunkSizeBytes)
-                return (fileSize, 1);
+                return ((int)fileSize, 1);
 
             int calculatedChunks = (int)Math.Ceiling((double)fileSize / _options.RecommendedChunkSizeBytes);
 
@@ -28,7 +28,7 @@ namespace FileService.Infrastructure.S3
 
             long chunkSize = (fileSize + actualChunks - 1) / actualChunks;
 
-            return (chunkSize, actualChunks);
+            return ((int)chunkSize, actualChunks);
         }
     }
 }
